+++ contracts/svm/programs/dummy/tests/helpers/helpers_express_relay.rs
@@ -1,4 +1,9 @@
 use {
+    crate::helpers::helpers_general::{
+        create_and_fund_kp,
+        create_and_submit_tx,
+        fund_pk,
+    },
     anchor_lang::{
         solana_program::sysvar::instructions as sysvar_instructions,
         InstructionData,
@@ -7,73 +12,16 @@ use {
     express_relay::sdk::test_helpers::create_initialize_express_relay_ix,
     solana_program_test::{
         anchor_processor,
-        BanksClientError,
         ProgramTest,
         ProgramTestContext,
     },
     solana_sdk::{
-        account::Account,
-        commitment_config::CommitmentLevel,
-        instruction::{
-            Instruction,
-            InstructionError::Custom,
-        },
+        instruction::Instruction,
         pubkey::Pubkey,
         signature::Keypair,
         signer::Signer,
-        transaction::{
-            Transaction,
-            TransactionError::{
-                self,
-                InstructionError,
-            },
-        },
     },
 };
-
-pub fn fund_pk(pk: Pubkey, program_test: &mut ProgramTest, balance: u64) {
-    program_test.add_account(
-        pk,
-        Account {
-            lamports: balance,
-            ..Account::default()
-        },
-    );
-}
-
-pub fn create_and_fund_kp(program_test: &mut ProgramTest, balance: u64) -> Keypair {
-    let kp = Keypair::new();
-    program_test.add_account(
-        kp.pubkey(),
-        Account {
-            lamports: balance,
-            ..Account::default()
-        },
-    );
-    kp
-}
-
-pub async fn create_and_submit_tx(
-    program_test_context: &mut ProgramTestContext,
-    ixs: &[Instruction],
-    payer: &Keypair,
-    signers: &[&Keypair],
-) -> Result<(), BanksClientError> {
-    let tx = Transaction::new_signed_with_payer(
-        ixs,
-        Some(&payer.pubkey()),
-        signers,
-        program_test_context
-            .get_new_latest_blockhash()
-            .await
-            .unwrap(),
-    );
-
-    program_test_context
-        .banks_client
-        .process_transaction_with_commitment(tx, CommitmentLevel::Processed)
-        .await
-}
 
 pub struct SetupInfo {
     pub program_test_context: ProgramTestContext,
@@ -134,20 +82,5 @@ pub fn create_do_nothing_ix(payer: Pubkey, permission: Pubkey, router: Pubkey) -
             router,
         }
         .to_account_metas(None),
-    }
-}
-
-pub fn assert_custom_error(error: TransactionError, instruction_index: u8, custom_error: u32) {
-    match error {
-        InstructionError(index, error_variant) => {
-            assert_eq!(index, instruction_index);
-            match error_variant {
-                Custom(code) => {
-                    assert_eq!(code, custom_error);
-                }
-                _ => panic!("Unexpected error code"),
-            }
-        }
-        _ => panic!("Unexpected error variant"),
     }
 }
